Better deletion integration
Several changes, but the most important one - if you delete the entire
list of source files in an earlier job, Trimmer marks each file as
<source file deleted> and disables the submit button.  If a source file
isn't listed (it's not there, or it's been recycled), the job processor
skips over it and sets that item to "(no source file)".  Looper also
adds a condition to the deleting function that doesn't pass non-files to
the recycling stage. Submit also restores it when you open another
.looper file.