Context menu tweak - added tsMuxeR detection back
Tweaked the new context menu code to once again detect if there is a .mts or .m2ts file in the selection, similar to the old handler.  If there is one, then the menu routine adds the tsMuxeR capable options to the "Transcoding" submenu - if there aren't, it just hides those options... sneakily.  Also added a little bit of commenting action in places.